Holon energy fixes
Alas twas not meant to be. But at least it wasn’t just GL that had bugs. Changed all holon energys to only check for basic energys of the required type. Added an EX check to the w portion of holon energy wp, and added more cases to remove special conditions with GL.