Added method isPokemonDelta
If this works as I intend, whenever a card would normally use .cardTypes.is(DELTA) instead it would call the created method and if ampheros delta is in play it would return true. I have very little understanding of how things work outside of implementation so I would greatly appreciate it if @axpendix could look it over and/or tell me what I did obviously wrong.