Probably the most stupid way to do this...
flagrama was right.... again.... Dude why do you have to be so good at spotting bugs in my code? The solution to this is to stop writing bad code but ill do anything to avoid personal responsibility. 
On a side note there has to be a better way to access the true card name and I asked admin about it but I think this will work as a proof of concept.