Some DP/MT fixes, Implemented MT 1-67 (#568)
* DP: Make Leech Seed last priority

* Debug print for Fossil Excavator (MT 111) and potential fix for crash-on-click-on-search

* Fix Retreat Cost for Pikachu (MT 94)

* Make Chesto Berry on Buizel (MT 75) prevent next coinflip

The coinflip for sleep is done regardless of it stopping to be asleep, no matter the outcome buizel already woke up.

* Partial Implementation of Vigoroth (MT 68)

May need a global ability for that sleep check, need to think about it.

* Have Magmortar LV.X (123) override burn dmg, not increase it

Also check for it not being higher already. This should be applied to other burn increasers like Infernape (UPR)

* MT: Implemented 55-67

* MT: Implemented 45-54

* Fix: Have Budew (DP 43) assert having deck before using Errand-Running

* MT: Implemented 40-44

* MT: Implemented 38-39

* Commented some uses of Source.POKEMONPOWER

Being handled as abilities for the time being. Will look around for adding these around later.

* Partially Implemented Unown I (MT 37)

Hidden Power in this one might be a bit of a nightmare, depending on the complexity of some checks...

* Implemented Toxicroak (MT 36)

* Remove debug code from Statics, do it directly in Fossil Excavator instead

Left a single change to the static, which I think is needed (and might solve the existing crash).

* Implemented MT 13-35, Fixed some Nidoran names

Nidoran name fixes on MT, RR, SW and TM

* Implemented MT 1-12

Some missing parts, but overall that'd cover most of the set.

* Fix "Heap Up" (Aggron MT 1)

* Fix "Psychic Guard" (Alakazam MT 2)

* Removed an unneeded assert (Ambipom MT 3)

* Fix "Flame Drum" (Magmortar MT 12)

* Potential Fix for static isValidFossilCard()?

Might be this, don't see any uses of bg without colons in the methods I'm aware work. Might be the cause for the crash?