Multiple SM card fixes (#814)
* Fix Charizard (TEU)

Knockouts are only checked after the whole ability is performed. So order of operations would be powerUsed->counters->Energy search and potential attach->KO check.

* Fix Naganadel & Guzzlord-GX (CEC)

Should be able to discard an UB even if not healing anything.

* Make Naganadel-GX able to discard an UB even without deck

Discard is an effect, not a cost unless stated otherwise (see: SSH Cinccino)

* Made Melmetal (UNB) able to discard a [M] Pokémon for no heal

Discard is an effect, not a cost in Abilities unless explicitly stated otherwise.

* Fixed Salazzle (UNB) so it doesn't assert deck

Roast Reveal has a ruling saying it can just discard an energy and do nothing else.

* Fixed Incineroar-GX (TEU)

Same fix as Charizard, shouldn't be knocked out until the full ability is done

* Fixed Mamoswine (CIN)

Works like Froslass (UNM) now. First choose the counters to put, then damage, then put the damage counters on it.