Minor fixes (#816)
* Update HeartgoldSoulsilver.groovy

* Update HolonPhantoms.groovy

I think this will now give the user the error message of "You haven't played a Supporter with Holon in its name this turn" when the player has not played a holon supporter but their discard pile has basic energy in it. This is more clear rather than giving the "You don't have the required energy in your discard" message when the player hasn't played a Holon Supporter anyway.