Fixed conditional switches from bench to active (#860)
* swFromBench: SSH-RCL

* swFromBench: UPR-LOT

* swFromBench: SUM-CIN, SMP, fix for FLI

* swFromBench: FRLG-DP

Missing some cards but this should good for now, covering all groovy cards.

* New sw2 static, removed swFromBench

sw gives a boolean regarding if it was successful or not, this should be a cleaner solution (also allow switch from bench when only given a single source).

* sw2: Gen 8

* sw2: Gen 7

* swFromBench: TEU-CEC

* sw2: Gen 3

Also 1 card from gen 4

* Remove unneeded targetting in some damage calls

* switchYourOpponentsBenchedWithActive()

Added a static for this. Could be more complex, but would rather use it for this simple case rn. checks bench again in case the attack is copied and the bench assert skipped.