"EX Dragon Frontiers" Fixes (#875)
* Minor changes for DF 2-4

Simplifying, moving conditions out of loops, adding showToOpponent() calls.

* Fix Nidoking Delta (DF 6)

EX series cards "Basic Pokémon or Evolution Card" is meant to be read as "Pokémon card". Discard on the attack should be doable even with no opp.bench

* Minor change on Nidoqueen Delta (DF 7)

* Fix Pinsir Delta (DF 9)

Should use self.owner.opposite as to avoid issues on opponent's turn.

* Fix Snorlax Delta (DF 10)

* Minor changes for DF 11-12, 15

* Changes in Gligar Delta (DF 16) and Mantine Delta (DF 20)

* Fix Quagsire Delta (DF 21) and Tropius Delta (DF 23)

* Quagsire should pick up at least one tool if the power is used. And it's once per turn, but not "1 power of that name per turn".

* Tropical Heal is not restricted to 1 power of that name per turn.

* Minor changes on Xatu Delta (DF 25)

Check for the attacker before looping.

* Minor changes for DF 26, 29-34

* Minor fixes in DF 39, 42, 49, 50, 59-60

* Minor fix on DF 63, Fixed Smoochum Delta (DF 64)

Using a single search now.

* Fixed Trapinch Delta (DF 69)

It should be swift damage, not counters.

* Added missing message to Island Hermit, fixed Mr. Stone's Project (DF 76-77)

* Minor logic change in Old Rod

Now prints a message when 2 heads/tails are flipped, but nothing else happens.

* Old Rod cannot be used if no valid card is in the discard

It is known it will fail. (Old rod itself won't be discarded until its effect is done)

* Minor changes in Altaria-ex (DF 90)

* Changed part of the logic in Dragonite ex Delta (DF 91)

Made the whole attack targeted, as to avoid loading the delayed when not needed.

Now unregisters after the first Knockout check. Should avoid potential issues with between-turn damages potentially triggering this before again.

* Minor change in Flygon ex Delta (DF 92), Fixed Kingdra ex Delta (DF 94)

Damage reduction should be applied to attacks from the opponent's Pokémon, not all. And the attachment should be post-damage.

* Minor changes in Latias-ex (DF 95)

Mostly simplifying the filter code on the Power, and adding an unregister for the KO check.

* Apply same simplification in Latios-ex Delta (DF 96)

* Minor fixes on Charizard Star Delta (DF 100)

* Fixed a trailing dot on the default Imprison/SW marker check

* Minor text fixes in Mew Star Delta (DF 101)