Refactor CardInfo getCollectionLineNo():int to getNumber():String
Some sets have card numbers that have string characters, such as promo sets. To be compatible with carddb, the field type needed to be String. It also renames the method getCollectionLineNo() to getNumber() to have compatible naming.

The scan url scheme is still zerofilled but that will also change after these changes are done.