Hidden fates fixes (#906)
* Add powerUsed to Jolteon Ability

* Fix Arbok's Last Pattern
Check that knockout was caused by damage from an attack before activating

* Fix Roxie+Koffing from CEC
Check if ability is blocked. Only check abilities if card being
checked is a PokemonCard

* Correct Weezing and Jessie & James
Have Jessie & James handle the activation of the ability like Roxie and Koffing
in CEC. Only the player of the card gets the effect. The effect cannot be used
disjoint from playing Jessie & James.

* Add hand checks to reduce number of empty selects