Several Card Fixes (#928)
* Fix Liepard (DAA)

Should now properly prevent it getting paralyzed.

* Make Flareon-GX (SM171)'s first attack effect optative

Can now be canceled at any point of energy selection.

* Solve an NPE related to Lurantis (SM25) and Virizion-GX (LOT)

Might need to be reworked later, but addAll solves the invisible PCS remaining in Virizion's place (causing the NPE with Lurantis)

* Fix a missing import in HGSS.groovy

* Mimikyu (GRI) now uses keyStore

It's also now linked to the player, not the card. So Marshadow-GX can now copy it properly.

* Correct Scorbunny (SWSH002) Super Singe's damage

* Revert Virizion-GX changes

This reverts commit feb0e82796706034c236dd2a74a61c1786f6b359 partially, since #921 seems to solve this situation. Leaving the check in Lurantis SM25 though, just in case.

* Set-related fixes

Changed the abbreviation for Sun&Moon-Base to SUM, following the official set code for it. Also added some EX tags in POP Series cards, and marked said exs as Rarity.RARE (POP sets have no holo cards).

* Make Nihilego-GX's ability source its SPC apply as ability-sourced

* Fix Salamence-GX (DRM)

Should only give your own non-EX/GX Pokémon free retreat

* Fix Primeape (TEU)

Should do extra damage, then discard the stadium if one is in play.

* Fixed Incineroar-GX (SUM)'s Hustling Strike damage output

Should only count benched [R] Pokémon now.

* Fix "Fight Song" in Altaria (DRM)

* Print the returned-to-deck item in the log for Florges (FLI)

* Also Show "Wondrous Gift" to the opponent

They might not be familiar with the card, or got discarded and they didn't get to see it earlier.

* Fix Electropower (LOT) ignoring swift attacks

Should be applied, as the boost is an effect on the attacking Pokémon, not on the defending one.

* Fix Gorebyss (CES)

Should reduce all damage, not only that less or equal than 40. Also simplified the dm() loop

* do gxCheck and gxPerform in Naganadel-GX (FLI)

* Make Ultra Necrozma-GX (FLI) discard after damage.

* Fix Metal Frying Pan (FLI)

Should only prevent damage done by opponent's attacks (so Steelix LOT should not reduce its self damage)

* Make Lucario (Promo SM54) point to the BUS print of the same card

Should fix Stance not working.

* Use afterDamage for Naganadel-GX's energy discard

* Update Charizard & Braixen-GX (CEC)

Should now offer to search, and force to take at least one card if chosen to do so.

* Fix Lucario-GX's Aura Strike attack

Make it only boost when evolved from a Riolu, and not from (e.g.) a Ditto Prism.

* Fix Dana (TEU)

Should now allow to search for up to 2 cards, and not reveal them in log.

* Fix Zebstrika (TEU)

Same fix as Lucario-GX earlier in the PR

* Fix Shaymin Prism Star (TEU)

Should now only count basic Energy cards.

* Add Weakness to Wobbuffet (BUS)

* Update Bonnie (FLI)

Based on a JP ruling, Bonnie won't discard a Prism Star stadium but it'll still enable the player's Zygarde-GX to use their GX attack during that turn.

Added some complementary bcs for this.

* Missing assert fail message for Bonnie (FLI)

* Fix Swellow (GRI)

Should now do a turn check and apply the damage boost on the correct attack.

* Add extra checks on Articuno (TEU)

Should only protect one's own Pokémon, and only benched ones at that.

* Fix Trumbeak (LOT)

Should not show the top card to its owner, only to the opponent. If it gets sent to the lost zone, the owner will see it in the log and in the lost zone directly.

Added a message for when the card is left where it was.

* Fix Mimikyu (GRI)

Should only work with non-GX attacks.

* Fix Hapu (UNM)

Should not show the opponent which cards were picked from the 6.

* Fix Dedenne (UNM)

Should do 20 damage, and only ask when you can draw cards.

* Make Pidgeot (TEU) only scoop the Active Pokémon

Should not be a choice from opp.all