UPR89: Don't show useless dialog to click through (#1192)
If there are no more M energy in the discard pile this card would
display the player's Pokémon to choose from that would have to be
clicked through to finish the attack. Instead, this will now just
skip any remaining checks if the discard pile no longer has any M
energy in it.