Add "usingThisAbilityEndsTurn" static, make use of it in DP cards with turn-ending Poké-Powers (#1273)
* Use "doWhenBlocked" in powerUsed(), add "usingThisAbilityEndsTurn" static

This will force abilities like Blastoise (SW) to end the player's turn, even if blocked by cards like Alakazam (MT) and Power Spray.

* Implement usingThisAbilityEndsTurn for the rest of DP cards

* Add an turn-ending message to Call Energy

* Use usingThisAbilityEndsTurn in Dialga Lv.X's Pokepower

This one's slightly different, as it only happens when not blocked, and two heads are flipped. But the static should work properly, and add a broadcast message.