Multiple Card Fixes (TR, Neo Genesis, Aquapolis, CG, DS, MD, UNM, DAA, RCL) (#1343)
* Fix Rocket's Sneak Attack (RO 72)

Should no longer crash if no trainers are found. Will just show the opponent's hand and finish.

* Fix Jynx (UNM 76)

Ominous Posture should not be useable when there's no bench (so no two pokemon to move counters into.

* Fix Retreat cost of Vaporeon (Majestic Dawn 34)

Reduced from 2 to 1, as it's meant to be.

* Fix Phione (Majestic Dawn 12)

Should not take damage from using its first attack.

* Fix Mawile (Crystal Guardians 9)

Should now check for the selected trainer being a tool before offering to attach it.

* Attempt at fixing Dust Island (Unbroken Bonds)

* Revert "Attempt at fixing Dust Island (Unbroken Bonds)"

This reverts commit 0228dc2efd0d800036cea473d4c7bc8eac15320a.

* Fix attempt for Skarmory (Delta Species 55)

Cry For Help should now properly filter in the selection popup. Also fixed a typo.

* Fix Milo (Rebel Clash 161)

The discard is a cost, not an effect. Thus, you need to have both cards in your deck and at least one card in hand to discard for Milo.

* Fix Dunsparce (Darkness Ablaze 137)

Should now properly check for the opponent's deck.