Fix Cyrus’s Conspiracy
Cyrus’s Conspiracy is able to search out any item when it should only be able to search out items that have Team Galactic’s Invention in their name. I would assume this is because the code says that it has to be either an item or Team Galactic’s Invention when it should be both.